moving ContextShift to cron package, as well as add some comments and name changes
